+++ commandline/Albatross.CommandLine.CodeGen/CommandOptionSetup.cs
@@ -1,6 +1,7 @@
 ﻿using Albatross.CodeAnalysis.Symbols;
 using Humanizer;
 using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
 using System;
 using System.Linq;
 
@@ -8,10 +9,10 @@ namespace Albatross.CommandLine.CodeGen {
 	public class CommandOptionSetup {
 		public CommandOptionSetup(IPropertySymbol property, AttributeData? optionAttribute) {
 			this.Property = property;
+			this.PropertyInitializer = GetPropertyInitializer(property);
 			this.OptionAttribute = optionAttribute;
 			this.Name = $"--{property.Name.Kebaberize()}";
 			this.Type = property.Type.ToDisplayString();
-			this.Required = property.Type.SpecialType != SpecialType.System_Boolean && !property.Type.IsNullable() && !property.Type.IsCollection();
 			this.Hidden = false;
 			this.Description = null;
 			this.Aliases = Array.Empty<string>();
@@ -24,6 +25,9 @@ namespace Albatross.CommandLine.CodeGen {
 				if (optionAttribute.TryGetNamedArgument("Required", out var required)) {
 					this.Required = Convert.ToBoolean(required.Value);
 				}
+				if (optionAttribute.TryGetNamedArgument("DefaultToInitializer", out var defaultToInitializer)) {
+					this.DefaultToInitializer = Convert.ToBoolean(defaultToInitializer.Value);
+				}
 				if (optionAttribute.TryGetNamedArgument("Hidden", out var hidden)) {
 					this.Hidden = Convert.ToBoolean(hidden.Value);
 				}
@@ -31,10 +35,13 @@ namespace Albatross.CommandLine.CodeGen {
 					this.Description = descriptionConstant.Value?.ToString();
 				}
 			}
+			this.Required = property.Type.SpecialType != SpecialType.System_Boolean && !property.Type.IsNullable() && !property.Type.IsCollection() && !ShouldDefaultToInitializer;
 		}
 
 		public string CommandOptionPropertyName => $"Option_{this.Property.Name}";
 		public IPropertySymbol Property { get; }
+		public ExpressionSyntax? PropertyInitializer { get; }
+		public bool DefaultToInitializer { get; }
 		public AttributeData? OptionAttribute { get; }
 		public bool Required { get; private set; }
 		public string Name { get; }
@@ -42,5 +49,16 @@ namespace Albatross.CommandLine.CodeGen {
 		public string? Description { get; private set; }
 		public bool Hidden { get; set; }
 		public string[] Aliases { get; }
+		public bool ShouldDefaultToInitializer => DefaultToInitializer && PropertyInitializer != null;
+
+		static ExpressionSyntax? GetPropertyInitializer(IPropertySymbol propertySymbol) {
+			foreach (var syntaxReference in propertySymbol.DeclaringSyntaxReferences) {
+				var syntaxNode = syntaxReference.GetSyntax();
+				if (syntaxNode is PropertyDeclarationSyntax propertyDeclaration) {
+					return propertyDeclaration.Initializer?.Value;
+				}
+			}
+			return null;
+		}
 	}
 }