+++ commandline/Sample.CommandLine/BaseCommandOptions.cs
@@ -1,6 +1,0 @@
-﻿namespace Sample.CommandLine {
-	public record class BaseCommandOptions {
-		public int Id { get; set; }
-		public string Name { get; set; } = string.Empty;
-	}
-}